requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
streamlit>=1.32.0
